// src/api/vcci-news/models/userUpdate.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * VietProDev CMS Backend API
 * Generated API documentation
 * OpenAPI spec version: 1.0.0
 */
import type { UserMutate } from './userMutate';
import type { UserUpdateType } from './userUpdateType';
import type { UserUpdateGender } from './userUpdateGender';

export type UserUpdate = UserMutate & {
  username?: string;
  first_name?: string;
  last_name?: string;
  phone?: string;
  avatar_id?: string;
  type?: UserUpdateType;
  birth_date?: string;
  /** @nullable */
  hometown?: string | null;
  gender?: UserUpdateGender;
  /** @nullable */
  bio?: string | null;
};